requirements.txt
flask>=2.0.1
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
pyserial>=3.5 